+++ rando/feedback/urls.py
@@ -1,0 +1,9 @@
+
+from django.conf.urls import patterns, url
+from .views import FeedBackView
+
+
+urlpatterns = patterns(
+    '',
+    url(r'feedback/$', FeedBackView.as_view(), name='feedback_view'),
+)