+++ rando/trekking/context_processors.py
@@ -15,4 +15,5 @@ def main(request):
 
         'PRINT_ENABLED': settings.PRINT_ENABLED,
         'VIEW3D_ENABLED': settings.VIEW3D_ENABLED,
+        'FEEDBACK_FORM_ENABLED': settings.FEEDBACK_FORM_ENABLED,
     }